requirements.txt
# Database requirements
pymysql>=1.0.2
cryptography>=35.0.0
sqlalchemy>=2.0.36

# project requirements
python-dotenv==1.0.0
pytest==7.4.0
astral==3.2
rpy2>=3.5.1
pytest>=6.2.0
rich>=13.9.4
dotmap==1.3.30
tqdm==4.67.1

# Analysis requirements
pandas>=2.2.3
numpy>=1.20.0
scikit-learn>=0.24.0
category_encoders>=2.6.4
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=5.4.0
networkx==3.4.2
fastparquet>=2024.11.0
statsforecast==2.0.0

# Optional but recommended
jupyter>=1.0.0
ipykernel>=6.0.0